refactor(server): extract GitHub webhook tests
